python/tracing/respan-tracing-sdk/pyproject.toml
[tool.poetry]
name = "example-scripts"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11,<3.14"
google-genai = ">=1.50.0,<2.0.0"
langgraph = "*"
langchain-anthropic = "*"
langchain-openai = "*"
typing-extensions = "*"
python-dotenv = "*"
respan-tracing = "*"
anthropic = "^0.86.0"
respan-instrumentation-openinference = "^1.0.1"
openinference-instrumentation-anthropic = "^1.0.0"
crewai = "^1.11.1"
openinference-instrumentation-crewai = "^1.1.0"
claude-agent-sdk = "^0.1.58"
openinference-instrumentation-claude-agent-sdk = "^0.1.0"
respan-ai = "^4.0.0"

[tool.poetry.group.dev.dependencies]
ipykernel = ">=7.1.0,<8.0.0"
instructor = "^1.13.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
